[#10] - refactor code according new requirements
