+++ src/main/java/com/catbytes/reviews/controller/rest/ProductController.java
@@ -1,9 +1,6 @@
 package com.catbytes.reviews.controller.rest;
 
-import com.catbytes.reviews.entity.Product;
 import com.catbytes.reviews.service.ProductService;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -16,10 +13,6 @@ public class ProductController {
         this.productService = productService;
     }
 
-    @PostMapping
-    public ResponseEntity<Product> addProduct(@RequestBody Product product) {
-        Product createdProduct = productService.addProduct(product);
-        return new ResponseEntity<>(createdProduct, HttpStatus.CREATED);
-    }
+    //TODO: get all products
 }
