+++ src/main/java/com/catbytes/reviews/mapper/ProductMapper.java
@@ -1,0 +1,10 @@
+package com.catbytes.reviews.mapper;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class ProductMapper {
+
+    //TODO: write implementation
+
+}
